+++ src/Message/UpdateCurrenciesMessage.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Message;
+
+final class UpdateCurrenciesMessage
+{
+    public function __construct(
+        public readonly array $codes = []
+    ) {}
+}